remove matrix_common.h
