+++ src/tyr/serializers.cc
@@ -251,7 +251,7 @@ std::string serializePbf(Api& request) {
 }
 
 // Generate leg shape in geojson format.
-baldr::json::MapPtr geojson_shape(const std::vector<PointLL> shape) {
+baldr::json::MapPtr geojson_shape(const std::vector<midgard::PointLL> shape) {
   auto geojson = baldr::json::map({});
   auto coords = baldr::json::array({});
   coords->reserve(shape.size());
