+++ src/tyr/trace_serializer.cc
@@ -10,7 +10,6 @@ using namespace valhalla;
 using namespace valhalla::midgard;
 using namespace valhalla::baldr;
 using namespace valhalla::odin;
-using namespace valhalla::thor;
 
 namespace {
