+++ test/matrix_bss.cc
@@ -14,7 +14,7 @@
 
 #include "sif/costfactory.h"
 #include "sif/dynamiccost.h"
-#include "thor/matrix_common.h"
+#include "thor/matrixalgorithm.h"
 #include "thor/timedistancebssmatrix.h"
 
 using namespace valhalla;
