+++ valhalla/thor/costmatrix.h
@@ -15,7 +15,6 @@
 #include <valhalla/sif/dynamiccost.h>
 #include <valhalla/sif/edgelabel.h>
 #include <valhalla/thor/edgestatus.h>
-#include <valhalla/thor/matrix_common.h>
 #include <valhalla/thor/matrixalgorithm.h>
 #include <valhalla/thor/pathinfo.h>
