+++ valhalla/thor/pathalgorithm.h
@@ -155,32 +155,32 @@ protected:
 
   // if `true` clean reserved memory for edge labels
   bool clear_reserved_memory_;
+};
 
-  /**
-   * Check for path completion along the same edge. Edge ID in question
-   * is along both an origin and destination and origin shows up at the
-   * beginning of the edge while the destination shows up at the end of
-   * the edge.
-   * @param  edgeid       Edge id.
-   * @param  origin       Origin path location information.
-   * @param  destination  Destination path location information.
-   */
-  virtual bool IsTrivial(const baldr::GraphId& edgeid,
-                         const valhalla::Location& origin,
-                         const valhalla::Location& destination) const {
-    for (const auto& destination_edge : destination.correlation().edges()) {
-      if (destination_edge.graph_id() == edgeid) {
-        for (const auto& origin_edge : origin.correlation().edges()) {
-          if (origin_edge.graph_id() == edgeid &&
-              origin_edge.percent_along() <= destination_edge.percent_along()) {
-            return true;
-          }
+/**
+ * Check for path completion along the same edge. Edge ID in question
+ * is along both an origin and destination and origin shows up at the
+ * beginning of the edge while the destination shows up at the end of
+ * the edge.
+ * @param  edgeid       Edge id.
+ * @param  origin       Origin path location information.
+ * @param  destination  Destination path location information.
+ */
+inline bool IsTrivial(const baldr::GraphId& edgeid,
+                      const valhalla::Location& origin,
+                      const valhalla::Location& destination) {
+  for (const auto& destination_edge : destination.correlation().edges()) {
+    if (destination_edge.graph_id() == edgeid) {
+      for (const auto& origin_edge : origin.correlation().edges()) {
+        if (origin_edge.graph_id() == edgeid &&
+            origin_edge.percent_along() <= destination_edge.percent_along()) {
+          return true;
         }
       }
     }
-    return false;
   }
-};
+  return false;
+}
 
 // Container for the data we iterate over in Expand* function
 struct EdgeMetadata {
