+++ valhalla/thor/timedistancebssmatrix.h
@@ -16,7 +16,6 @@
 #include <valhalla/sif/edgelabel.h>
 #include <valhalla/thor/astarheuristic.h>
 #include <valhalla/thor/edgestatus.h>
-#include <valhalla/thor/matrix_common.h>
 #include <valhalla/thor/matrixalgorithm.h>
 
 namespace valhalla {
