+++ valhalla/thor/timedistancematrix.h
@@ -14,7 +14,6 @@
 #include <valhalla/sif/dynamiccost.h>
 #include <valhalla/sif/edgelabel.h>
 #include <valhalla/thor/edgestatus.h>
-#include <valhalla/thor/matrix_common.h>
 #include <valhalla/thor/matrixalgorithm.h>
 
 namespace valhalla {
