+++ valhalla/tyr/serializers.h
@@ -15,8 +15,8 @@
 #include <valhalla/baldr/rapidjson_utils.h>
 #include <valhalla/meili/match_result.h>
 #include <valhalla/midgard/gridded_data.h>
+#include <valhalla/midgard/pointll.h>
 #include <valhalla/proto/api.pb.h>
-#include <valhalla/thor/matrix_common.h>
 #include <valhalla/tyr/actor.h>
 
 namespace valhalla {
@@ -129,7 +129,7 @@ baldr::json::ArrayPtr serializeWarnings(const valhalla::Api& api);
  * @param shape  The points making up the line.
  * @returns The GeoJSON geometry of the LineString
  */
-baldr::json::MapPtr geojson_shape(const std::vector<PointLL> shape);
+baldr::json::MapPtr geojson_shape(const std::vector<midgard::PointLL> shape);
 
 // Elevation serialization support
